[PO22] Initialize repo (#99)
Here we are yet again! 
I literally just copy pasted and tried to understand last years' unit
testing. I got it working, although JUnit is way too verbose on errors,
I don't know if that is meant to happen. I also updated junit to version
1.9.0. I commented out the reflection bits as I am not sure how Tagus
paper is this year and I wanted to keep it simple so people can
contribute asap.

Here is an example of JUnit logs on failed tests, it's doube printing
the same information, I don't know if this normal. Let me know and thank
you for the repo, tests and help in advance!


![junit1](https://user-images.githubusercontent.com/82618983/197014664-1434675e-a165-47e6-be00-6f1dbeb7f536.png)


![junit2](https://user-images.githubusercontent.com/82618983/197014638-1b23b31f-a8a3-4043-93c4-309c12059339.png)